pyproject.toml
[project]
name = "mmda"
version = "0.0.1"
authors = [
  { name="Po-han Li", email="pohanli@utexas.edu" },
]
description = "Text and Image Feature Extraction"
readme = "README.md"
requires-python = ">=3.10.4"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]  # list of folders that contain the packages (["."] by default)
include = ["mmda"]  # package names should match these glob patterns (["*"] by default)
exclude = []  # exclude packages matching these glob patterns (empty by default)
namespaces = false  # to disable scanning PEP 420 namespaces (true by default)

[tool.poetry]
name = "mmda"
version = "0.0.1"
authors = ["Po-han Li <pohanli@utexas.edu>"]
packages = [{include = "mmda"}]
description = "Multi-modal Data Alignment"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.13"
numpy = "*"
matplotlib = "*"
seaborn = "*"
pandas = "*"
argparse = "*"
flake8 = "*"
black = "*"
scikit-learn = "*"
hydra-core = "*"
hydra-joblib-launcher = "*"
protobuf = "3.20.*"
torchvision = "*"
torchaudio = "*"
ipython = "^8.15.0"
scipy = "^1.11.2"
swarm-visualizer = {git = "https://github.com/UTAustin-SwarmLab/Swarm-Visualization"}
cca-zoo = "^2.5.0"
ipykernel = "^6.29.3"
jupyter = "^1.0.0"
gpustat = "*"
datasets = "^2.18.0"
# LLaVA = {git = "https://github.com/haotian-liu/LLaVA.git"}
torch = "2.1.1"
transformers = "^4.13.0"
soundfile = "*"
librosa = "*"
wget = "*"
wavfile = "*"
resampy = "*"
open-clip-torch = "2.24.0"
sentence-transformers = "2.4.0"
yt-dlp = "*"

[project.optional-dependencies]
dev = ["black", "ruff", "mypy"]
test = ["pytest", "pytest-cov", "pytest-mock"]

[tool.black]
line-length = 120

[tool.ruff]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "UP",  # pyupgrade
  "D",   # pydocstyle
]
line-length = 120

ignore = ["ANN101", "ANN102", "D100"]

extend-exclude = [".venv", "venv", "vscode"]

[tool.ruff.pydocstyle]
convention = "google"